ml-service/requirements.txt
# Web framework
Flask==2.3.3
flask-cors==4.0.0

# ML dependencies - CPU-only versions
torch==2.7.0 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.22.0 --index-url https://download.pytorch.org/whl/cpu

# CLIP model
open_clip_torch==2.20.0

# Image processing
Pillow==9.5.0


# Scientific computing
numpy==1.24.3

# HTTP requests
requests==2.31.0

# Computer vision
opencv-python-headless==4.8.1.78

# Model utilities
timm==0.9.12

# OCR for character recognition
easyocr==1.7.1
opencv-python==4.8.1.78